feat: Update evaluation panel and related components for improved progress tracking and error handling

+++ src/aios/cli/eval_cli.py
@@ -44,7 +44,7 @@ def eval_run(
         "-b",
         help="Batch size or 'auto'",
     ),
-    limit: Optional[int] = typer.Option(
+    limit: Optional[float] = typer.Option(
         None,
         "--limit",
         "-l",
